Deleted follow up intent on last follow up intent of domain deletion

+++ api/modules/intent/controllers/deleteById.intent.controller.js
@@ -1,6 +1,7 @@
 'use strict';
 const Async = require('async');
 const Boom = require('boom');
+const _ = require('lodash')
 const IntentTools = require('../tools');
 const Status = require('../../../helpers/status.json');
 
@@ -10,6 +11,7 @@ module.exports = (request, reply) => {
     const server = request.server;
     const redis = server.app.redis;
     let followUpIntents;
+    let followUpDomainToDelete = false;
 
     const getScenario = (intentId) => new Promise((resolve, reject) => {
         // let followUpIntents = [];
@@ -29,12 +31,12 @@ module.exports = (request, reply) => {
 
         });
     });
-    const intentsToDelete = [];
+    let intentsToDelete = [];
 
     function getTreeRecursively(id) {
         return getScenario(id, server).then(function (intent) {
             let node = {id};
-            if (!intent.followUpIntents){
+            if (!intent.followUpIntents) {
                 node.children = [];
                 return Promise.resolve(node);
             }
@@ -46,6 +48,9 @@ module.exports = (request, reply) => {
         });
     }
 
+    let parentIntentId;
+    let parentIntentScenarioData;
+
     Async.waterfall([
             (cb) => {
                 getTreeRecursively(intentId, server).then((tree) => {
@@ -55,10 +60,16 @@ module.exports = (request, reply) => {
             (cb) => {
                 getScenario(intentId).then((scenario) => {
                     if (scenario.parentIntent >= 0) {
+                        parentIntentId = scenario.parentIntent;
                         getScenario(scenario.parentIntent).then((parentIntentScenario) => {
+                            parentIntentScenarioData = _.cloneDeep(parentIntentScenario);
                             parentIntentScenario.followUpIntents = parentIntentScenario.followUpIntents.filter((item) => {
                                 return item !== intentId;
                             });
+                            if (parentIntentScenario.followUpIntents.length === 0) {
+                                // No more follow up intent in follow up domain
+                                followUpDomainToDelete = true;
+                            }
                             delete parentIntentScenario.id;
                             delete parentIntentScenario.agent;
                             delete parentIntentScenario.domain;
@@ -81,8 +92,69 @@ module.exports = (request, reply) => {
                         return cb(null);
                 })
             },
+            (callbackDeleteDomain) => {
+                if (!followUpDomainToDelete)
+                    return callbackDeleteDomain(null)
+                let domainId;
+                let agentId;
+
+                Async.series([
+                    (callbackGetAgent) => {
+                        redis.zscore(`agents`, parentIntentScenarioData.agent, (err, score) => {
+
+                            if (err) {
+                                const error = Boom.badImplementation(`An error occurred retrieving the agent id of the agent ${parentIntentScenarioData.agent}`);
+                                return callbackGetAgent(error);
+                            }
+                            agentId = score;
+                            return callbackGetAgent(null);
+                        });
+                    },
+                    (callbackGetDomain) => {
+
+                        redis.zscore(`agentDomains:${agentId}`, 'FollowUp-' + parentIntentId, (err, score) => {
+
+                            if (err) {
+                                const error = Boom.badImplementation(`An error occurred retrieving the id of the domain ${'FollowUp-' + parentIntentId}`);
+                                return callbackGetDomain(error);
+                            }
+                            domainId = score;
+                            return callbackGetDomain(null);
+                        });
+                    },
+                    (cb) => {
+                        let options = {
+                            method: 'DELETE',
+                            url: `/domain/${domainId}`,
+                        }
+                        server.inject(options, (res) => {
+                            if (res.statusCode !== 200) {
+                                if (res.statusCode === 404) {
+                                    const errorNotFound = Boom.notFound('The specified domain doesn\'t exists');
+                                    return cb(errorNotFound);
+                                }
+                                const error = Boom.create(res.statusCode, `An error occurred deleting the domain ${'FollowUp-' + parentIntentId}`);
+                                return cb(error, null);
+
+                            }
+                            return cb(null);
+                        });
+                    }], (err, result) => {
+                    if (err) {
+                        return callbackDeleteDomain(err);
+                    }
+                    else {
+                        //domain has been deleted since last follow up intent has been deleted
+                        intentsToDelete = [];
+                        return callbackDeleteDomain(null);
+                    }
+                })
+
+            },
 
             (cb) => {
+                if (intentsToDelete.length === 0)
+                    return cb(null,{message:'successful operation'})
                 Async.forEach(intentsToDelete, (followUpIntent, callback) => {
 
                     IntentTools.deleteIntentTool(server, redis, followUpIntent, (err, result) => {
